+++ webui/src/App.tsx
@@ -2,6 +2,7 @@ import { Navigate, Route, Routes } from "react-router-dom";
 import { AppShell } from "./components/AppShell";
 import { CandidatesPage } from "./pages/CandidatesPage";
 import { ComparePage } from "./pages/ComparePage";
+import { DailyReviewPage } from "./pages/DailyReviewPage";
 import { MarketDataPage } from "./pages/MarketDataPage";
 import { PortfolioPage } from "./pages/PortfolioPage";
 import { ResearchPage } from "./pages/ResearchPage";
@@ -15,6 +16,7 @@ export default function App() {
       <Route element={<AppShell />}>
         <Route index element={<Navigate to="/trends" replace />} />
         <Route path="trends" element={<TrendDiscoveryPage />} />
+        <Route path="review" element={<DailyReviewPage />} />
         <Route path="screen" element={<ScreenWorkbench />} />
         <Route path="candidates" element={<CandidatesPage />} />
         <Route path="compare" element={<ComparePage />} />